+++ example/android/app/src/main/kotlin/dev/mixin27/app_update_manager_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package dev.mixin27.app_update_manager_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()